Update ModifyModal.Stories.tsx
feat: added story in the modify modal
